common: Add DONGLE ERROR result
Sometimes the PTS dongle jams or there is a problem with the power supply.
After that we got series of INONC results, not caught by recovery.
Then in PTS logs there are info about non-pass verdict from APICOM or
stack init fail. We can filter them and return DONGLE ERROR instead of
INCOC and run recovery.