ptscontrol: Force restarts of BPV in long runs
Workaround for Bluetooth Protocol Viewer leaks. It is visible in
the Process Explorer that after each test case 2 new process handles
to PTS.exe are generated and accumulate over time to thousands. It may
be a symptom that some resources are not being freed and maybe this
is a cause of some Windows system crashes and restarts at long runs.